06-lab-complete/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
redis>=5.0.0
openai>=1.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
pytest>=7.4.0
